fixup: Addressed comments
