+++ pslogin/src/main/scala/WorldSessionActor.scala
@@ -148,7 +148,7 @@ class WorldSessionActor extends Actor with MDCContextAware {
     case PlayerStateMessageUpstream(_) =>
     
     case msg @ ChatMsg(messagetype, unk1, recipient, contents) =>
-      log.info("Handling " + msg)
+      log.info("Chat: " + msg)
       
       // TODO: Depending on messagetype, may need to prepend sender's name to contents with proper spacing
       // TODO: Just replays the packet straight back to sender; actually needs to be routed to recipients!
